+++ R/tab.R
@@ -7,9 +7,12 @@
 #' }
 #'
 #' @param .data Expression to append, typically pulled from the pipe \code{ \%>\% }
+#' @param inverse Invert match behavior, defaults to \code{FALSE} (match
+#' tabs). Use \code{FALSE} to \emph{not} match tabs.
 #'
 #' @examples
 #' rx_tab()
+#' rx_tab(inverse = TRUE)
 #'
 #' # create an expression
 #' x <- rx_tab()
@@ -20,6 +23,10 @@
 #' # extract match
 #' regmatches(string, regexpr(x, string))
 #' @export
-rx_tab <- function(.data = NULL) {
-  paste0(.data, "\\t") # shouldn't this just be \t
+rx_tab <- function(.data = NULL, inverse = FALSE) {
+  switch(as.character(inverse),
+         "FALSE" = paste0(.data, "\\t"),
+         "TRUE" = paste0(.data, "[^\\t]"),
+         stop("Inverse accepts either TRUE (don't match tabs) or FALSE (default, match tabs)")
+  )
 }