+++ tests/testthat/test-tab.R
@@ -10,4 +10,14 @@ test_that("tab special character works", {
   # expect match
   expect_true(grepl(rx_tab(), "hello\tworld"))
 
+  # expect inverse behavior
+  expect_false(grepl(rx_tab(inverse = TRUE), "\\t"))
+
+  # expect matching carriage returns and newlines
+  expect_true(grepl(rx_tab(inverse = TRUE), "\\r"))
+  expect_true(grepl(rx_tab(inverse = TRUE), "\\n"))
+
+  # expect error if incorrect value to inverse arg
+  expect_error(rx_tab(inverse = 1))
+
 })