+++ tests/testthat/test-whitespace.R
@@ -16,4 +16,13 @@ test_that("whitespace special character works", {
   # dont expect match
   expect_false(grepl(rx_whitespace(), "abc"))
 
+  # expect inverse behavior
+  expect_false(grepl(rx_whitespace(inverse = TRUE), " "))
+  expect_false(grepl(rx_whitespace(inverse = TRUE), "\t"))
+  expect_false(grepl(rx_whitespace(inverse = TRUE), "\n"))
+  expect_false(grepl(rx_whitespace(inverse = TRUE), "\r"))
+
+  # expect error if incorrect value to inverse arg
+  expect_error(rx_whitespace(inverse = 1))
+
 })